refactor [#95] User groups API 호출 함수를 `user` feature에서 관리
