+++ src/features/user/query/use-user-query.tsx
@@ -1,5 +1,5 @@
 import { useQuery } from "@tanstack/react-query";
-import { getUser } from "../apis";
+import { getUser, getUserGroups } from "../apis";
 
 export function useUserQuery() {
   const { data } = useQuery({
@@ -8,4 +8,13 @@ export function useUserQuery() {
   });
 
   return { user: data };
+}
+
+export function useUserGroupsQuery() {
+  const { data } = useQuery({
+    queryKey: ["user", "groups"],
+    queryFn: getUserGroups,
+  });
+
+  return { userGroups: data };
 }
